Add files via upload
.

📌 Day 5 – Type Casting in Java
🔹 What is Type Casting?

Type casting in Java is the process of converting one data type into another. It helps us to use values in different forms when needed.

For example, sometimes we need to change a smaller data type into a bigger one (like int to double), or a bigger one into a smaller one (like double to int).